add sirius service
